refactored vehicle forms
