+++ package.json
@@ -9,6 +9,7 @@
     "bootstrap": "^4.3.1",
     "firebase": "^7.2.1",
     "formik": "^2.0.4",
+    "lodash.clonedeep": "^4.5.0",
     "moment": "^2.24.0",
     "moment-range": "^4.0.2",
     "react": "^16.11.0",
